fix(workflow): apply Bolt 2 review (never-raise key, DTO layering)
Rebased onto the #320-fixed foundations head. Relocate Bolt 2 runtime DTOs
(ReturnAck/StepOutputRecord/WorkflowIndexRow/StepState) to a light
workflow_runtime module so the MCP server no longer pulls jsonschema/yaml.
Add non-string-key never-raise regression tests; chain delete OSError cause.
Refs #312
